+++ src/index.ts
@@ -1,6 +1,7 @@
 /* eslint-disable no-console */
 import { createApp } from './app.js'
 import {
+  community,
   groupsToLeave,
   groupToJoin,
   isBehindProxy,
@@ -8,8 +9,9 @@ import {
   webId,
 } from './config/index.js'
 
-createApp({ webId, isBehindProxy, groupToJoin, groupsToLeave }).then(app =>
-  app.listen(port, async () => {
-    console.log(`community inbox service is listening on port ${port}`)
-  }),
+createApp({ webId, isBehindProxy, groupToJoin, groupsToLeave, community }).then(
+  app =>
+    app.listen(port, async () => {
+      console.log(`community inbox service is listening on port ${port}`)
+    }),
 )