+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Allow providing community or group URI as activity object
+
 ## [0.2.1] - 2025-01-07
 
 ### Fixed